+++ lib/index.ts
@@ -1,12 +1,12 @@
-import { DeltaChat } from './deltachat'
+import { AccountManager } from './deltachat'
 
-export default DeltaChat
+export default AccountManager
 
 export { Chat } from './chat'
 export { ChatList } from './chatlist'
 export { C } from './constants'
 export { Contact } from './contact'
-export { DeltaChat }
+export { AccountManager as DeltaChat }
 export { Locations } from './locations'
 export { Lot } from './lot'
 export { Message, MessageState, MessageViewType } from './message'